refactor(plms): rename esm/ -> esmc/, delete esm2hf/ (ferritin-uhd)
- Rename src/esm/ to src/esmc/ to clearly identify it as the
  EvolutionaryScale ESM-C/ESM-3 module (not Facebook ESM-2)
- Update all internal crate::esm:: paths to crate::esmc::
- Update lib.rs: pub mod esm -> pub mod esmc, re-export paths
- Delete esm2hf/ (Python-only stub, zero Rust content, not tracked)
- Add module-level doc comments to esmc/mod.rs and esm2/mod.rs
  distinguishing model families, HuggingFace repos, and license status

cargo check and cargo test -p ferritin-plms pass.
