+++ .beads/interactions.jsonl
@@ -1,0 +1,5 @@
+{"id":"int-d7ef0428","kind":"field_change","created_at":"2026-05-26T00:18:52.192804Z","actor":"Zachary Charlop-Powers","issue_id":"ferritin-3nr","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Closed"}}
+{"id":"int-9ccfc10b","kind":"field_change","created_at":"2026-05-26T00:18:52.507613Z","actor":"Zachary Charlop-Powers","issue_id":"ferritin-uvi","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Closed"}}
+{"id":"int-32357f1b","kind":"field_change","created_at":"2026-05-26T00:18:52.79116Z","actor":"Zachary Charlop-Powers","issue_id":"ferritin-4k6","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Closed"}}
+{"id":"int-eb2e73c0","kind":"field_change","created_at":"2026-05-26T00:42:40.777411Z","actor":"Zachary Charlop-Powers","issue_id":"ferritin-mhn","extra":{"field":"status","new_value":"closed","old_value":"open","reason":"All 3 child issues completed: embedding scale fix, token dropout, gelu_erf. Epic fully delivered."}}
+{"id":"int-60a3eacc","kind":"field_change","created_at":"2026-05-26T00:45:09.756401Z","actor":"Zachary Charlop-Powers","issue_id":"ferritin-uhd","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Completed: deleted esm2hf/, renamed esm/ -> esmc/, updated all crate::esm:: paths to crate::esmc::, updated lib.rs re-exports, added module-level doc comments to both esmc/mod.rs and esm2/mod.rs distinguishing the two model families. cargo check and cargo test -p ferritin-plms pass."}}
