--- crates/ferritin-plms/src/esm/mod.rs
@@ -1,6 +1,0 @@
-pub mod layers;
-pub mod models;
-// pub mod pretrained;
-// pub mod sdk;
-pub mod tokenization;
-pub mod utils;
