+++ crates/ferritin-plms/src/esmc/layers/attention.rs
@@ -1,5 +1,5 @@
-use crate::esm::layers::rotary::RotaryEmbedding;
-use crate::esm::models::esmc::ESMCConfig;
+use crate::esmc::layers::rotary::RotaryEmbedding;
+use crate::esmc::models::esmc::ESMCConfig;
 use candle_core::{Module, Result};
 use candle_nn::{self as nn, LayerNormConfig, VarBuilder};
 // use scaled_dot_product_attention;
