+++ crates/ferritin-plms/src/esmc/layers/codebook.rs
