+++ crates/ferritin-plms/src/esmc/tokenization/sequence_tokenizer.rs
@@ -1,4 +1,4 @@
-use crate::esm::utils::constants::esm3::SEQUENCE_VOCAB;
+use crate::esmc::utils::constants::esm3::SEQUENCE_VOCAB;
 use anyhow::Result;
 use std::collections::HashMap;
 use std::sync::Arc;
