+++ crates/ferritin-plms/src/esmc/utils/constants/esm3.rs
