+++ crates/ferritin-plms/src/esmc/utils/decoding.rs
