+++ crates/ferritin-plms/src/esmc/utils/encoding.rs
