+++ crates/ferritin-plms/src/esmc/utils/sampling.rs
