+++ crates/ferritin-plms/src/esmc/utils/structure/affine3D.rs
