+++ crates/ferritin-plms/src/esmc/utils/structure/protein_chain.rs
