+++ crates/ferritin-plms/src/esmc/utils/structure/protein_complex.rs
