+++ crates/ferritin-plms/src/esmc/utils/structure/protein_structure.rs
